# Replace usages of `cargoLock` with `cargoHash`