--- a/pkgs/cosmic-ext-observatory/package.nix
+++ b/pkgs/cosmic-ext-observatory/package.nix
@@ -25,23 +25,8 @@ rustPlatform.buildRustPackage {
     hash = "sha256-r3YBdDEh2EfPp4XKhSCwQj4VUHg89q9vWpm53O46Fes=";
   };
 
-  cargoLock = {
-    lockFile = ./Cargo.lock;
-    outputHashes = {
-      "accesskit-0.16.0" = "sha256-yeBzocXxuvHmuPGMRebbsYSKSvN+8sUsmaSKlQDpW4w=";
-      "atomicwrites-0.4.2" = "sha256-QZSuGPrJXh+svMeFWqAXoqZQxLq/WfIiamqvjJNVhxA=";
-      "clipboard_macos-0.1.0" = "sha256-tovB4fjPVVRY8LKn5albMzskFQ+1W5ul4jT5RXx9gKE=";
-      "cosmic-client-toolkit-0.1.0" = "sha256-/DJ/PfqnZHB6VeRi7HXWp0Vruk+jWBe+VCLPpiJeEv4=";
-      "cosmic-config-0.1.0" = "sha256-q/n0vdbSSRne0gNtNCM3IvcVNy6VS2/vli+q7Dnmjsw=";
-      "cosmic-freedesktop-icons-0.2.6" = "sha256-+GvdoGHlTzu3joGkX89c+4lsc/l6FkbAqN4TD9IO8t8=";
-      "cosmic-text-0.12.1" = "sha256-nCw3RNIHINXH4+m9wKB+0CeoXSVKKxP+ylaZhfp8u+o=";
-      "dpi-0.1.1" = "sha256-whi05/2vc3s5eAJTZ9TzVfGQ/EnfPr0S4PZZmbiYio0=";
-      "iced_glyphon-0.6.0" = "sha256-u1vnsOjP8npQ57NNSikotuHxpi4Mp/rV9038vAgCsfQ=";
-      "smithay-clipboard-0.8.0" = "sha256-4InFXm0ahrqFrtNLeqIuE3yeOpxKZJZx+Bc0yQDtv34=";
-      "softbuffer-0.4.1" = "sha256-a0bUFz6O8CWRweNt/OxTvflnPYwO5nm6vsyc/WcXyNg=";
-      "taffy-0.3.11" = "sha256-SCx9GEIJjWdoNVyq+RZAGn0N71qraKZxf9ZWhvyzLaI=";
-    };
-  };
+  useFetchCargoVendor = true;
+  cargoHash = "sha256-4BnulyS0qQBTh1tl5uS/TAgu3Onlg3Au5RvfI32n9dM=";
 
   nvtop = fetchFromGitHub {
     owner = "Syllo";