--- a/pkgs/cosmic-notifications/package.nix
+++ b/pkgs/cosmic-notifications/package.nix
@@ -20,26 +20,8 @@ rustPlatform.buildRustPackage {
     hash = "sha256-9/uApnu8KyC+7DNu5L6J006K6CtNjRHFS7gOJATGwQU=";
   };
 
-  cargoLock = {
-    lockFile = ./Cargo.lock;
-    outputHashes = {
-      "accesskit-0.16.0" = "sha256-yeBzocXxuvHmuPGMRebbsYSKSvN+8sUsmaSKlQDpW4w=";
-      "atomicwrites-0.4.2" = "sha256-QZSuGPrJXh+svMeFWqAXoqZQxLq/WfIiamqvjJNVhxA=";
-      "clipboard_macos-0.1.0" = "sha256-tovB4fjPVVRY8LKn5albMzskFQ+1W5ul4jT5RXx9gKE=";
-      "cosmic-client-toolkit-0.1.0" = "sha256-/DJ/PfqnZHB6VeRi7HXWp0Vruk+jWBe+VCLPpiJeEv4=";
-      "cosmic-config-0.1.0" = "sha256-q/n0vdbSSRne0gNtNCM3IvcVNy6VS2/vli+q7Dnmjsw=";
-      "cosmic-freedesktop-icons-0.2.6" = "sha256-+GvdoGHlTzu3joGkX89c+4lsc/l6FkbAqN4TD9IO8t8=";
-      "cosmic-panel-config-0.1.0" = "sha256-XbuTvUWS3Z3FMTSm1SYcGxxa5T6FGT2AQQ9uzXH+PlI=";
-      "cosmic-settings-daemon-0.1.0" = "sha256-9Pq5WFBeIRvP2VZaa3BzoqiQmzN6taa20u7k+2aF3v0=";
-      "cosmic-text-0.12.1" = "sha256-nCw3RNIHINXH4+m9wKB+0CeoXSVKKxP+ylaZhfp8u+o=";
-      "cosmic-time-0.4.0" = "sha256-Hz1l915bVNrgjNS+dMBZlLSdtS/RIMX+U5xcsjah26A=";
-      "dpi-0.1.1" = "sha256-whi05/2vc3s5eAJTZ9TzVfGQ/EnfPr0S4PZZmbiYio0=";
-      "iced_glyphon-0.6.0" = "sha256-u1vnsOjP8npQ57NNSikotuHxpi4Mp/rV9038vAgCsfQ=";
-      "smithay-clipboard-0.8.0" = "sha256-4InFXm0ahrqFrtNLeqIuE3yeOpxKZJZx+Bc0yQDtv34=";
-      "softbuffer-0.4.1" = "sha256-a0bUFz6O8CWRweNt/OxTvflnPYwO5nm6vsyc/WcXyNg=";
-      "taffy-0.3.11" = "sha256-SCx9GEIJjWdoNVyq+RZAGn0N71qraKZxf9ZWhvyzLaI=";
-    };
-  };
+  useFetchCargoVendor = true;
+  cargoHash = "sha256-h5OPeK8WpFuBRXcnQDrgT9zO7f5pPynIJNVoMCqpNwc=";
 
   nativeBuildInputs = [
     libcosmicAppHook