While investigating my system flake's evaluation time I noticed that removing nixos-cosmic caused a considerable speedup—through some searching I found https://github.com/NixOS/nixpkgs/issues/327063, which led me to believe that the usage of `cargoLock` across the packages in this repository was the culprit.

After converting all of the packages to use `cargoHash` via the script linked in https://github.com/NixOS/nixpkgs/issues/357257, my evaluation time dropped by 6 seconds, as evidenced by `hyperfine`:

Before:
```
Benchmark 1: nix eval '.#nixosConfigurations."desktop".config.system.build.toplevel' --no-eval-cache --impure
  Time (mean ± σ):     17.984 s ±  0.130 s    [User: 13.300 s, System: 1.888 s]
  Range (min … max):   17.869 s … 18.126 s    3 runs
```

After:
```
Benchmark 1: nix eval '.#nixosConfigurations."desktop".config.system.build.toplevel' --no-eval-cache --impure
  Time (mean ± σ):     12.115 s ±  0.115 s    [User: 9.270 s, System: 1.286 s]
  Range (min … max):   12.013 s … 12.239 s    3 runs
```

I ran `nix build` on a couple of the packages to make sure nothing was obviously broken, and they built fine, but I lack the processing power to check all of them 😅

Also, I am not sure if any changes would have to be made to the CI or Cachix action as a result of this, it [looks like](https://github.com/lilyinstarlight/nixos-cosmic/blob/9502f7621771096901a83e8a2c05e978dfa1644f/pkgs/cosmic-screenshot/package.nix#L21) they can handle `cargoHash` though!