@drakon64 2024-12-15
Currently the new fetchCargo vendor requires https://github.com/NixOS/nixpkgs/pull/282798 to be merged otherwise a lot of builds are going to fail.

Also, fetchCargo isn't available in 24.11 yet. I'm pretty sure there was a PR to backport it though.